+++ cmake/FleCSI/kokkos.cmake
@@ -1,4 +1,4 @@
 macro(flecsi_enable_kokkos target)
-  find_package(Kokkos REQUIRED)
+  find_package(Kokkos COMPONENTS separable_compilation REQUIRED)
   target_link_libraries(${target} PUBLIC Kokkos::kokkos)
 endmacro()