+++ flecsi/flog/packet.hh
@@ -17,7 +17,6 @@
 #endif
 
 #include <sys/time.h>
-#include <unistd.h>
 
 #include <algorithm>
 #include <array>
@@ -27,15 +26,6 @@
 #include <sstream>
 #include <thread>
 #include <vector>
-
-#ifndef FLOG_MAX_PACKET_BUFFER
-#define FLOG_MAX_PACKET_BUFFER 1024
-#endif
-
-// Microsecond interval
-#ifndef FLOG_PACKET_FLUSH_INTERVAL
-#define FLOG_PACKET_FLUSH_INTERVAL 100000
-#endif
 
 /// \cond core
 namespace flecsi {