+++ flecsi/run/backend.hh
@@ -26,11 +26,12 @@
 #endif // FLECSI_BACKEND
 
 namespace flecsi::run {
+inline std::optional<dependencies_guard> dependent;
 // Now that the backend's context_t is complete:
+inline std::optional<context_t> context::ctx;
 context_t &
 context::instance() {
-  static context_t context;
-  return context;
+  return *ctx;
 } // instance
 } // namespace flecsi::run