+++ flecsi/run/context.cc
@@ -4,10 +4,9 @@ namespace po = boost::program_options;
 
 namespace flecsi::run {
 
-struct context::backend_arg {
-  static auto & args() {
-    return instance().backend_args_;
-  }
+namespace {
+struct backend_arg {
+  static inline arguments::argv * argv;
   static const std::string & get(const std::vector<std::string> & v) {
     flog_assert(v.size() == 1, "wrong token count");
     return v.front();
@@ -15,7 +14,7 @@ struct context::backend_arg {
   struct single {
     friend void
     validate(boost::any &, const std::vector<std::string> & sv, single *, int) {
-      args().push_back(get(sv));
+      argv->push_back(get(sv));
     }
   };
   struct space {
@@ -23,45 +22,42 @@ struct context::backend_arg {
     validate(boost::any &, const std::vector<std::string> & sv, space *, int) {
       using I = std::istream_iterator<std::string>;
       std::istringstream iss(get(sv));
-      args().insert(args().end(), I(iss), I());
+      argv->insert(argv->end(), I(iss), I());
     }
   };
 };
+} // namespace
 
-int
-context::initialize_generic(int argc, char ** argv, bool dependent) {
-  if(const auto p = std::getenv("FLECSI_SLEEP")) {
-    const auto n = std::atoi(p);
-    std::cerr << getpid() << ": sleeping for " << n << " seconds...\n";
-    sleep(n);
-  }
+arguments::arguments(int argc, char ** argv) {
+  act.program = argv[0] ? argv[0] : "";
+  act.program = act.program.substr(act.program.rfind('/') + 1);
+#ifdef FLECSI_ENABLE_MPI
+  dep.mpi.push_back(act.program);
+#endif
+#ifdef FLECSI_ENABLE_KOKKOS
+  dep.kokkos.push_back(act.program);
+#endif
+  act.code = getopt(argc, argv); // also populates act.stderr
+}
 
-  initialize_dependent_ = dependent;
-
-  // Save command-line arguments
-  for(auto i(0); i < argc; ++i) {
-    argv_.push_back(argv[i]);
-  } // for
-
-  program_ = argv[0];
-  program_ = program_.substr(program_.rfind('/') + 1);
-
+status
+arguments::getopt(int argc, char ** argv) {
   po::options_description master("Basic Options");
   master.add_options()("help,h", "Print this message and exit.");
 
   // Add externally-defined descriptions to the main description
-  for(auto & od : descriptions_map_) {
+  auto & dm = context::descriptions_map();
+  for(auto & od : dm) {
     if(od.first != "FleCSI Options") {
       master.add(od.second);
     } // if
   } // for
 
   po::options_description flecsi_desc =
-    descriptions_map_.count("FleCSI Options")
-      ? descriptions_map_["FleCSI Options"]
-      : po::options_description("FleCSI Options");
+    dm.count("FleCSI Options") ? dm["FleCSI Options"]
+                               : po::options_description("FleCSI Options");
 
-  // Together, these initialize backend_args_.
+  backend_arg::argv = &cfg.backend;
   flecsi_desc.add_options()("backend-args",
     po::value<backend_arg::space>(),
     "Pass arguments to the backend. The single argument is a quoted "
@@ -73,8 +69,6 @@ context::initialize_generic(int argc, char ** argv, bool dependent) {
     "multiple times.");
 #if defined(FLECSI_ENABLE_FLOG)
   std::string flog_tags_;
-  int flog_verbose_;
-  int64_t flog_output_process_;
   // Add FleCSI options
   flecsi_desc.add_options() // clang-format off
       (
@@ -87,7 +81,7 @@ context::initialize_generic(int argc, char ** argv, bool dependent) {
       )
       (
         "flog-verbose",
-        po::value(&flog_verbose_)
+        po::value(&cfg.flog.verbose)
           ->implicit_value(1)
           ->default_value(0),
         "Enable verbose output. Passing '-1' will strip any additional"
@@ -95,7 +89,7 @@ context::initialize_generic(int argc, char ** argv, bool dependent) {
       )
       (
         "flog-process",
-        po::value(&flog_output_process_)->default_value(0),
+        po::value(&cfg.flog.process)->default_value(0),
         "Restrict output to the specified process id. The default is process 0."
         " Use '--flog-process=-1' to enable all processes."
       ); // clang-format on
@@ -106,93 +100,95 @@ context::initialize_generic(int argc, char ** argv, bool dependent) {
   po::options_description all("All Options");
   all.add(master);
   all.add(flecsi_desc);
-  all.add(hidden_options_);
+  all.add(context::hidden_options());
 
+  auto & pd = context::positional_description();
   po::parsed_options parsed = po::command_line_parser(argc, argv)
                                 .options(all)
-                                .positional(positional_desc_)
+                                .positional(pd)
                                 .allow_unregistered()
                                 .run();
 
-  auto print_usage =
-    [this](std::string program, auto const & master, auto const & flecsi) {
-      if(process_ == 0) {
-        std::cout << "Usage: " << program << " ";
+  struct guard : std::ostringstream {
+    guard(std::string & s) : out(s) {}
+    ~guard() {
+      out = std::move(*this).str();
+    }
+    std::string & out;
+  } stderr(act.stderr);
+  auto usage = [&] {
+    stderr << "Usage: " << act.program << " ";
 
-        size_t positional_count = positional_desc_.max_total_count();
-        size_t max_label_chars = std::numeric_limits<size_t>::min();
+    size_t positional_count = pd.max_total_count();
+    size_t max_label_chars = std::numeric_limits<size_t>::min();
 
-        for(size_t i{0}; i < positional_count; ++i) {
-          std::cout << "<" << positional_desc_.name_for_position(i) << "> ";
+    for(size_t i{0}; i < positional_count; ++i) {
+      stderr << "<" << pd.name_for_position(i) << "> ";
 
-          const size_t size = positional_desc_.name_for_position(i).size();
-          max_label_chars = size > max_label_chars ? size : max_label_chars;
-        } // for
+      const size_t size = pd.name_for_position(i).size();
+      max_label_chars = size > max_label_chars ? size : max_label_chars;
+    } // for
 
-        max_label_chars += 2;
+    max_label_chars += 2;
 
-        std::cout << std::endl << std::endl;
+    stderr << "\n\n";
 
-        if(positional_count) {
-          std::cout << "Positional Options:" << std::endl;
+    if(positional_count) {
+      stderr << "Positional Options:" << std::endl;
 
-          for(size_t i{0}; i < positional_desc_.max_total_count(); ++i) {
-            auto const & name = positional_desc_.name_for_position(i);
-            auto help = positional_help_.at(name);
-            std::cout << "  " << name << " ";
-            std::cout << std::string(max_label_chars - name.size() - 2, ' ');
+      for(size_t i{0}; i < pd.max_total_count(); ++i) {
+        auto const & name = pd.name_for_position(i);
+        auto help = context::positional_help().at(name);
+        stderr << "  " << name << " ";
+        stderr << std::string(max_label_chars - name.size() - 2, ' ');
 
-            if(help.size() > 78 - max_label_chars) {
-              std::string first = help.substr(
-                0, help.substr(0, 78 - max_label_chars).find_last_of(' '));
-              std::cout << first << std::endl;
-              help =
-                help.substr(first.size() + 1, help.size() - first.size() + 1);
+        if(help.size() > 78 - max_label_chars) {
+          std::string first = help.substr(
+            0, help.substr(0, 78 - max_label_chars).find_last_of(' '));
+          stderr << first << std::endl;
+          help = help.substr(first.size() + 1, help.size() - first.size() + 1);
 
-              while(help.size() > 78 - max_label_chars) {
-                std::string part = help.substr(
-                  0, help.substr(0, 78 - max_label_chars).find_last_of(' '));
-                std::cout << std::string(max_label_chars + 1, ' ') << part
-                          << std::endl;
-                help = help.substr(part.size() + 1, help.size() - part.size());
-              } // while
+          while(help.size() > 78 - max_label_chars) {
+            std::string part = help.substr(
+              0, help.substr(0, 78 - max_label_chars).find_last_of(' '));
+            stderr << std::string(max_label_chars + 1, ' ') << part
+                   << std::endl;
+            help = help.substr(part.size() + 1, help.size() - part.size());
+          } // while
 
-              std::cout << std::string(max_label_chars + 1, ' ') << help
-                        << std::endl;
-            }
-            else {
-              std::cout << help << std::endl;
-            } // if
-
-          } // for
-
-          std::cout << std::endl;
+          stderr << std::string(max_label_chars + 1, ' ') << help << std::endl;
+        }
+        else {
+          stderr << help << std::endl;
         } // if
 
-        std::cout << master << std::endl;
-        std::cout << flecsi << std::endl;
+      } // for
+
+      stderr << std::endl;
+    } // if
+
+    stderr << master << std::endl;
+    stderr << flecsi_desc << std::endl;
 
 #if defined(FLECSI_ENABLE_FLOG)
-        auto const & tm = flog::state::instance().tag_map();
+    auto const & tm = flog::state::tag_map();
 
-        if(tm.size()) {
-          std::cout << "Available FLOG Tags (FleCSI Logging Utility):"
-                    << std::endl;
-        } // if
+    if(tm.size()) {
+      stderr << "Available FLOG Tags (FleCSI Logging Utility):" << std::endl;
+    } // if
 
-        for(auto t : tm) {
-          std::cout << "  " << t.first << std::endl;
-        } // for
+    for(auto t : tm) {
+      stderr << "  " << t.first << std::endl;
+    } // for
 #endif
-      } // if
-    }; // print_usage
+  };
 
   try {
     po::variables_map vm;
     po::store(parsed, vm);
 
     if(vm.count("help")) {
-      print_usage(program_, master, flecsi_desc);
+      usage();
       return status::help;
     } // if
     if(vm.count("control-model")) {
@@ -204,25 +200,33 @@ context::initialize_generic(int argc, char ** argv, bool dependent) {
 
     po::notify(vm);
 
+#ifdef FLECSI_ENABLE_FLOG
+    if(flog_tags_ != "none") {
+      std::istringstream is(flog_tags_);
+      std::string tag;
+      while(std::getline(is, tag, ','))
+        cfg.flog.tags.push_back(tag);
+    }
+#endif
+
     // Call option check methods
+    auto & oc = context::option_checks();
     for(auto const & [name, boost_any] : vm) {
-      auto const & ita = option_checks_.find(name);
-      if(ita != option_checks_.end()) {
+      auto const & ita = oc.find(name);
+      if(ita != oc.end()) {
         auto [positional, check] = ita->second;
 
         std::stringstream ss;
         if(!check(boost_any.value(), ss)) {
-          if(process_ == 0) {
-            std::string dash = positional ? "" : "--";
-            std::cerr << FLOG_COLOR_LTRED << "ERROR: " << FLOG_COLOR_RED
-                      << "invalid argument for '" << dash << name
-                      << "' option!!!" << std::endl
-                      << FLOG_COLOR_LTRED << (ss.str().empty() ? "" : " => ")
-                      << ss.str() << FLOG_COLOR_PLAIN << std::endl
-                      << std::endl;
-          } // if
+          std::string dash = positional ? "" : "--";
+          stderr << FLOG_COLOR_LTRED << "ERROR: " << FLOG_COLOR_RED
+                 << "invalid argument for '" << dash << name << "' option!!!"
+                 << std::endl
+                 << FLOG_COLOR_LTRED << (ss.str().empty() ? "" : " => ")
+                 << ss.str() << FLOG_COLOR_PLAIN << std::endl
+                 << std::endl;
 
-          print_usage(program_, master, flecsi_desc);
+          usage();
           return status::help;
         } // if
       } // if
@@ -236,24 +240,15 @@ context::initialize_generic(int argc, char ** argv, bool dependent) {
       error.replace(pos, 2, "");
     } // if
 
-    std::cerr << FLOG_COLOR_LTRED << "ERROR: " << FLOG_COLOR_RED << error
-              << "!!!" << FLOG_COLOR_PLAIN << std::endl
-              << std::endl;
-    print_usage(program_, master, flecsi_desc);
+    stderr << FLOG_COLOR_LTRED << "ERROR: " << FLOG_COLOR_RED << error << "!!!"
+           << FLOG_COLOR_PLAIN << std::endl
+           << std::endl;
+    usage();
     return status::command_line_error;
   } // try
 
-  unrecognized_options_ =
+  unrecognized =
     po::collect_unrecognized(parsed.options, po::include_positional);
-
-#if defined(FLECSI_ENABLE_FLOG)
-  if(flog::state::instance().initialize(
-       flog_tags_, flog_verbose_, flog_output_process_)) {
-    return status::error;
-  } // if
-#endif
-
-  initialized_ = true;
 
   return status::success;
 }