+++ doc/sphinx/src/news.rst
@@ -127,6 +127,7 @@ Fixed
 
   * Certain control-flow graphs compile with Graphviz support and are drawn correctly.
   * ``--backend-args`` can be specified multiple times. The collected arguments are passed to the backend.
+  * MPI and Kokkos are initialized with no arguments (so that they cannot misinterpret arguments not meant for them).
 
 * Data
 
@@ -157,7 +158,9 @@ Fixed
 
 * Logging
 
+  * Tags work correctly in tasks executed in parallel.
   * Messages are sorted by timestamp correctly.
+  * Exiting the process before ``flecsi::finalize`` does not crash.
 
 Changes in v2.1.0 (April 16 2021)
 +++++++++++++++++++++++++++++++++